+++ NEWS.md
@@ -1,5 +1,7 @@
 # testthat (development version)
 
+* Links generate for snapshot hints now work when using parallel tests (#1802).
+
 * Experimental `is_snapshotting()` to determine if code is running inside a 
   snapshot test (#1796).